requirements.txt
openai
langchain_community
langchain-openai
chromadb
langchain
flask
pandas
python-dotenv

